pyproject.toml
[tool.poetry]
name = "cdcrtool"
version = "0.1.0"
description = "Cross Document Coreference Ground Truthing Tool"
authors = ["James Ravenscroft <ravenscroftj@gmail.com>"]
license = "MIT"

packages = [
    {include = "cdcrapp"}
]

[tool.poetry.dependencies]
python = "^3.6.1"
streamlit = "^0.57.3"
pandas = "^1.0.3"
numpy = "^1.18.2"
sqlalchemy = "^1.3.16"
redis = "^3.4.1"
psycopg2-binary = "^2.8.5"
click = "^7.1.1"
alembic = "^1.4.2"
python-dotenv = "^0.13.0"
google-api-python-client = "^1.8.0"
google-auth-httplib2 = "^0.0.3"
google-auth-oauthlib = "^0.4.1"
tqdm = "^4.45.0"
newspaper3k = "^0.2.8"
scikit-learn = "^0.22.2"
transformers = "^2.9.0"
matplotlib = "^3.2.1"
seaborn = "^0.10.1"
flask = "^1.1.2"
flask-security = "^3.0.0"
email_validator = "^1.1.1"
bcrypt = "^3.1.7"
Flask-Markdown = "^0.3"
flask-restful = "^0.3.8"
flask-cors = "^3.0.8"
flask-login = "0.4.1"
neptune-client = "^0.4.119"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
autopep8 = "^1.5.3"
rope = "^0.17.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
